Fix remove/add browser problem
Actual for Linux/Windows. When a browser is shown in a JTabbedPane, and panes are switched, the component in a pane if first removed and then added back. This causes the browser canvas window to be disposed along with its child - the browser window itself. The problem has already been described and fixed as https://bitbucket.org/chromiumembedded/java-cef/issues/57. The solution was to temporarily re-parent the browser window to a hidden frame so that not to loose its context for the render process. However, the fix doesn't solve the problem with concurrency b/w EDT and Toolkit (main) threads. Whereas the re-parenting is asynchronously delegated to Toolkit, the disposal is initiated on EDT (on Linux right from EDT, on Windows synchronously from EDT to Toolkit). So, the disposal can happen before the re-parenting.

On Linux we additionally have to use XSync due to the X Window client-server async communication. So XReparentWindow should make a round-trip to X-server and back so that XDestroyWindow had actual window state.